+++ tests/caas/test_cluster_order_lifecycle.py
@@ -4,10 +4,10 @@ from pathlib import Path
 
 import pytest
 
-from tests.osac_cli import OsacCLI
-from tests.grpc_client import GRPCClient
-from tests.helpers import wait_for_cluster_deletion, wait_for_cluster_order_cr, wait_for_cluster_ready
-from tests.k8s_client import K8sClient
+from tests.core.grpc_client import GRPCClient
+from tests.core.helpers import wait_for_cluster_deletion, wait_for_cluster_order_cr, wait_for_cluster_ready
+from tests.core.k8s_client import K8sClient
+from tests.core.osac_cli import OsacCLI
 
 
 @pytest.fixture
